All comment mutation locked down.

+++ src/entity/comment.js
@@ -1,23 +1,19 @@
 import { GraphQLYogaError } from '@graphql-yoga/node';
 import prisma from '../prisma';
+import getUserId from '../utils/getUserId';
+import verifyComment from '../utils/verifyComment';
 
 export default {
-  createComment: async (_, { data }, { prismaSelect, pubsub }, info) => {
+  createComment: async (_, { data }, { prismaSelect, pubsub, request }, info) => {
+    const userId = getUserId(request);
     const select = prismaSelect(info);
-    const user = await prisma.user.findUnique({ where: { id: data.author } });
-    if (!user) throw new GraphQLYogaError('User not found!');
-
-    const post = await prisma.post.findUnique({ where: { id: data.post } });
-    if (!post) throw new GraphQLYogaError('Post not found!');
-
-    if (!post.published) throw new GraphQLYogaError('Post not published!');
 
     const comment = await prisma.comment.create({
       data: {
         text: data.text,
         author: {
           connect: {
-            id: data.author
+            id: userId
           }
         },
         post: {
@@ -38,8 +34,14 @@ export default {
 
     return comment;
   },
-  deleteComment: async (_, args, { prismaSelect, pubsub }, info) => {
+
+  deleteComment: async (_, args, { prismaSelect, pubsub, request }, info) => {
     try {
+      const userId = getUserId(request);
+      if (!(await verifyComment(args.id, userId))) {
+        throw new GraphQLYogaError('Unable to delete');
+      }
+
       const select = prismaSelect(info);
       const comment = await prisma.comment.delete({ where: { id: args.id }, ...select });
       pubsub.publish(`comment ${comment.postId}`, {
@@ -51,10 +53,14 @@ export default {
       return comment;
     } catch (error) {
       if (error.code === 'P2025') throw new GraphQLYogaError('Comment not found.');
-      throw new GraphQLYogaError('Error4!!');
+      throw new GraphQLYogaError(error);
     }
   },
-  updateComment: async (_, { id, data }, { prismaSelect, pubsub }, info) => {
+  updateComment: async (_, { id, data }, { prismaSelect, pubsub, request }, info) => {
+    const userId = getUserId(request);
+    if (!(await verifyComment(id, userId))) {
+      throw new GraphQLYogaError('Unable to update');
+    }
     const select = prismaSelect(info);
     const comment = await prisma.comment.update({ where: { id: id }, data, ...select });
